fix: move fixUrl and allowedOptions to utils
